+++ src/features/Request/Request.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Request = () => {
+  return <div>Some forms</div>
+}
+
+export default Request
