+++ src/features/RequestMaterials/RequestMaterials.jsx
@@ -1,130 +1,32 @@
 import React from 'react'
-import { Form, Select, Input, Button, Tooltip, Icon } from 'antd'
-import { shape, func } from 'prop-types'
-import * as shortId from 'short-id'
+import { Row, Tabs } from 'antd'
 
-import {
-  selectMaterials,
-  selectLvl,
-  fullNameInfo,
-  lvlInfo,
-  selectInfo,
-  noteInfo,
-  submitBtn,
-  FullName,
-  lvlUrgency,
-  itAccessories,
-  Note,
-} from 'features/constants'
+import MaterialIts from 'features/MaterialsIts/MaterialIts'
+import Request from 'features/Request/Request'
 
-const { Option } = Select
-
-const RequestMaterials = ({ form }) => {
-  const handleSubmit = e => {
-    e.preventDefault()
-    form.validateFields((err, values) => {
-      if (!err) {
-        localStorage.setItem('Request Materials', JSON.stringify(values))
-      }
-    })
-  }
-  const { getFieldDecorator } = form
-  const { TextArea } = Input
+const { TabPane } = Tabs
+const RequestMaterials = () => {
   return (
-    <Form
-      style={{
-        position: 'absolute',
-        top: '50%',
-        left: '55%',
-        transform: 'translate(-50%, -50%)',
-        width: '50%',
-      }}
-      labelCol={{ span: 5 }}
-      wrapperCol={{ span: 12 }}
-      onSubmit={handleSubmit}
-    >
-      <Form.Item
-        label={
-          <span>
-            {FullName}
-            &nbsp;
-            <Tooltip title={fullNameInfo}>
-              <Icon type="question-circle-o" />
-            </Tooltip>
-          </span>
-        }
-      >
-        {getFieldDecorator(FullName, {
-          rules: [
-            {
-              required: true,
-              message: fullNameInfo,
-              whitespace: true,
-            },
-          ],
-        })(<Input />)}
-      </Form.Item>
-      <Form.Item label="Select" hasFeedback>
-        {getFieldDecorator(lvlUrgency, {
-          rules: [{ required: true, message: lvlInfo }],
-        })(
-          <Select placeholder={lvlInfo}>
-            {selectLvl.map(lvl => (
-              <Option key={shortId.generate()} value={lvl}>
-                {lvl}
-              </Option>
-            ))}
-          </Select>,
-        )}
-      </Form.Item>
-      <Form.Item label="Select[multiple]">
-        {getFieldDecorator(itAccessories, {
-          rules: [
-            {
-              required: true,
-              message: selectInfo,
-              type: 'array',
-            },
-          ],
-        })(
-          <Select mode="multiple" placeholder={selectInfo}>
-            {selectMaterials.map(material => (
-              <Option key={shortId.generate()} value={material}>
-                {material}
-              </Option>
-            ))}
-          </Select>,
-        )}
-      </Form.Item>
-      <Form.Item
-        label={
-          <span>
-            Note&nbsp;
-            <Tooltip title={noteInfo}>
-              <Icon type="question-circle-o" />
-            </Tooltip>
-          </span>
-        }
-      >
-        {getFieldDecorator(Note, {
-          rules: [{ required: false, message: { noteInfo } }],
-        })(<TextArea />)}
-      </Form.Item>
-      <Form.Item wrapperCol={{ span: 12, offset: 5 }}>
-        <Button type="primary" htmlType="submit">
-          {submitBtn}
-        </Button>
-      </Form.Item>
-    </Form>
+    <>
+      <Row type="flex" justify="center">
+        <Tabs
+          style={{
+            width: '50%',
+            float: 'center',
+          }}
+          defaultActiveKey="1"
+          size="large"
+        >
+          <TabPane tab="Tab 1" key="1">
+            <MaterialIts />
+          </TabPane>
+          <TabPane tab="Tab 2" key="2">
+            <Request />
+          </TabPane>
+        </Tabs>
+      </Row>
+    </>
   )
 }
 
-RequestMaterials.propTypes = {
-  form: shape({
-    getFieldsValue: func,
-  }).isRequired,
-}
-
-const withForm = Form.create({ name: 'requestMaterials' })
-
-export default withForm(RequestMaterials)
+export default RequestMaterials
